[FEAT/#289] 출석체크 api 연결

+++ app/src/main/java/com/spoony/spoony/data/dto/response/SpoonDrawResponseDto.kt
@@ -25,7 +25,7 @@ data class SpoonDrawResponseDto(
         @SerialName("spoonAmount")
         val spoonAmount: Int,
         @SerialName("probability")
-        val probability: Int,
+        val probability: Double,
         @SerialName("spoonImage")
         val spoonImage: String
     )
